# [Performance] Character pet bulk saving