Similar to https://github.com/EQEmu/Server/pull/3336

When character `Save()` is triggered character pet data get saved individually for every buff, inventory item and info. In situations where character `Save()` routines are being triggered more aggressively, the query volume adds up quickly.

Attached is an example on Lazarus within 12 seconds there are almost 1k in queries with `Save()` related queries making up a bulk.

This PR reduces all individual insert queries into one bulk insert for each buff pet `info` `buff` and `inventory` row.

We also now only purge data from the table only if there is data to save

https://gist.github.com/Akkadius/0553f9fd49e8da9caf57ce6e0da4a475

```
  Zone |   Query    | QueryDatabase DELETE FROM character_pet_info WHERE char_id = 1 -- (2 rows affected) (0.000301s)-- [poair] (The Plane of Air) inst_id [0]
  Zone |   Query    | QueryDatabase INSERT INTO character_pet_info (char_id, pet, petname, petpower, spell_id, hp, mana, size, taunting)  VALUES (1,0,'Gebn000',0,635,2193811,0,5.000000,0),(1,1,'',0,0,0,0,0.000000,0) -- (2 rows affected) (0.000240s)-- [poair] (The Plane of Air) inst_id [0]
  Zone |   Query    | QueryDatabase DELETE FROM character_pet_buffs WHERE char_id = 1 -- (2 rows affected) (0.000217s)-- [poair] (The Plane of Air) inst_id [0]
  Zone |   Query    | QueryDatabase INSERT INTO character_pet_buffs (char_id, pet, slot, spell_id, caster_level, castername, ticsremaining, counters, numhits, rune, instrument_mod)  VALUES (1,0,0,6824,90,'',-1,0,0,0,10),(1,0,1,171,90,'',109,0,0,0,10) -- (2 rows affected) (0.000198s)-- [poair] (The Plane of Air) inst_id [0]
  Zone |   Query    | QueryDatabase DELETE FROM character_pet_inventory WHERE char_id = 1 -- (1 row affected) (0.000260s)-- [poair] (The Plane of Air) inst_id [0]
  Zone |   Query    | QueryDatabase INSERT INTO character_pet_inventory (char_id, pet, slot, item_id)  VALUES (1,0,13,111512) -- (1 row affected) (0.000238s)-- [poair] (The Plane of Air) inst_id [0]
```